Use +512 for 0x43 and +256 for 0x44 expander pins

+++ src/components/digitalIO/Wippersnapper_DigitalGPIO.h
@@ -20,7 +20,7 @@
 
 /** Holds data about a digital input pin */
 struct digitalInputPin {
-  uint8_t pinName; ///< Pin name
+  uint16_t pinName; ///< Pin name
   long period;     ///< Timer interval, in millis, -1 if disabled.
   long prvPeriod;  ///< When timer was previously serviced, in millis
   int prvPinVal;   ///< Previous pin value
@@ -42,14 +42,13 @@ public:
 
   void
   initDigitalPin(wippersnapper_pin_v1_ConfigurePinRequest_Direction direction,
-                 uint8_t pinName, float period,
+                 uint16_t pinName, float period,
                  wippersnapper_pin_v1_ConfigurePinRequest_Pull pull);
   void
   deinitDigitalPin(wippersnapper_pin_v1_ConfigurePinRequest_Direction direction,
-                   uint8_t pinName);
-
+                   uint16_t pinName);
   int digitalReadSvc(int pinName);
-  void digitalWriteSvc(uint8_t pinName, int pinValue);
+  void digitalWriteSvc(uint16_t pinName, int pinValue);
 #if defined(ARDUINO_ARDUINO_NESSO_N1)
   // void
   // initDigitalPin(wippersnapper_pin_v1_ConfigurePinRequest_Direction
